Implement function q_delete_dup
The input queue is already sorted, so removing duplicate nodes only
requires checking whether two consecutive nodes have the same string.
To achieve this efficiently, list_for_each_entry_safe is used,
as it allows iterating while keeping track of the next node.
The implementation removes the previous node upon detecting
a duplicate and records the next node. This ensures that the
last occurrence of a duplicate is also removed, preventing
cases where the final duplicate remains in the queue.

Reference:
https://leetcode.com/problems/remove-duplicates-from-sorted-list-ii/

Change-Id: I63e52c5d06eca120a34bbebd75c069469b873f6d