Implement function q_swap
First, the 'first' pointer is assigned to the first node and
the 'second' pointer is assigned to the second node in the queue.
The loop then begins to perform the swap process using 'list_move'.
After each swap, the 'first' and 'second' pointers are reassigned
to the next pair of adjacent nodes. It's important to note that
the loop exits when either 'first' or 'second' reaches the 'head',
which accounts for both odd and even length queues.

Reference:
https://leetcode.com/problems/swap-nodes-in-pairs/

Change-Id: Icb07d5833243bf3027dc8a544520f9928548d138