Implement function q_sort
Implement a recursive top-down 'merge sort' for sorting the elements
in the queue. The 'mergeSort' function recursively splits the linked
list into two halves until each part contains a single element.
Then, the 'merge' function is used to combine these parts back
together while maintaining the sorted order.
The 'q_sort' function uses 'mergeSort' to sort the queue based on the
provided 'descend' parameter, which determines whether the sorting
is in ascending or descending order.
In case of descending order, the result is reversed at the end.

Additionally, due to an error, the dummy list initialization
in 'merge' was modified to use 'LIST_HEAD'
instead of 'INIT_LIST_HEAD'.

Change-Id: Ic37fe69317a4ff5896253ffe71acd5152afcb47b